Move services to sub-packages
