+++ app/src/main/java/de/jordsand/birdcensus/activities/AreaSelectionMap.java
@@ -23,7 +23,7 @@ import de.jordsand.birdcensus.R;
 import de.jordsand.birdcensus.core.MonitoringArea;
 import de.jordsand.birdcensus.database.BirdCountOpenHandler;
 import de.jordsand.birdcensus.database.repositories.SQLiteMonitoringAreaRepository;
-import de.jordsand.birdcensus.services.SimpleBirdCountService;
+import de.jordsand.birdcensus.services.census.SimpleBirdCountService;
 
 /**
  * Shows the user a list of all available monitoring areas
