+++ app/src/main/java/de/jordsand/birdcensus/services/maps/OsmDroidOfflineMapSetupService.java
@@ -1,4 +1,4 @@
-package de.jordsand.birdcensus.services;
+package de.jordsand.birdcensus.services.maps;
 
 import java.io.File;
 import java.io.FileInputStream;
